Physics improvements
Wind has been improved upon heavily with better directional control in the editor. Still tweaking interactions between wind and the umbrella, but now depending on the orientation of the umbrella relative to the wind, the force is changed. So if the wind comes from behind the umbrella the force is greater than from the side, for example.